+++ src/layers/DTLayer.hpp
@@ -1,8 +1,9 @@
 #pragma once
 
 #include <Geode/Geode.hpp>
-#include "../../managers/StatsManager.hpp"
-#include "../../utils/Save.hpp"
+#include "Geode/ui/TextArea.hpp"
+#include "../managers/StatsManager.hpp"
+#include "../utils/Save.hpp"
 
 using namespace geode::prelude;