+++ src/layers/LabelSettingsLayer.hpp
@@ -1,8 +1,9 @@
 #pragma once
 
 #include <Geode/Geode.hpp>
-#include "../LabelLayoutWindow.hpp"
-#include "../cells/LabelFontCell.hpp"
+#include "../layers/LabelLayoutWindow.hpp"
+#include <Geode/ui/TextArea.hpp>
+#include "../layers/LabelFontCell.hpp"
 #include "../layers/DTLayer.hpp"
 
 using namespace geode::prelude;