+++ src/layers/LinkLevelCell.hpp
@@ -1,7 +1,8 @@
 #pragma once
 
 #include <Geode/Geode.hpp>
-#include "../../managers/StatsManager.hpp"
+#include "../managers/StatsManager.hpp"
+#include "../layers/DTLinkLayer.hpp"
 
 using namespace cocos2d;
 using namespace geode;