+++ src/managers/StatsManager.hpp
@@ -166,6 +166,7 @@ typedef struct {
     std::string type;
 } prismSetting;
 
+// matjson fuckery
 template <>
 struct matjson::Serialize<prismSetting> {
     static Result<prismSetting> fromJson(const matjson::Value& value) {