+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v2.4.4
+- added playtime calculation! (use special keys {ptf0},{ptrun} and {ptall} to see it!!)
+- changed default layout! (if you never changed your layout before, resetting it to default is recommended!)
+
 ## v2.4.3
 - organized code!!!!! added comments to everything!!!!!!